feat(hot-post): 인기글 점수 산정 방식을 로그 스케일로 개선
선형 가중합 대신 LOG(1 + count) 형태의 로그 스케일을 적용해
아웃라이어 게시글이 상위권을 독점하는 문제를 완화한다.

Before: count * weight
After:  LOG(1 + count) * weight

HotPostScoreView.getScore() 반환 타입을 long → double로 변경하고
관련 테스트의 점수 검증을 isCloseTo(within(0.001))로 업데이트한다.

Constraint: MySQL LOG() 함수(자연로그) 사용 — Write/Read DB 미분리 환경
Rejected: LOG2/LOG10 | 자연로그가 아웃라이어 완화 효과가 더 완만함
Rejected: 시간 감쇠(time decay) | 집계 기간(windowDays)으로 이미 제어 중
Confidence: high
Scope-risk: narrow

Resolves: #67

+++ src/test/java/kr/ac/knu/comit/post/service/PostServiceTest.java
@@ -434,7 +434,7 @@ class PostServiceTest {
         }
     }
 
-    private PostRepository.HotPostScoreView hotPostScore(Long postId, long score) {
+    private PostRepository.HotPostScoreView hotPostScore(Long postId, double score) {
         return new PostRepository.HotPostScoreView() {
             @Override
             public Long getPostId() {
@@ -442,7 +442,7 @@ class PostServiceTest {
             }
 
             @Override
-            public long getScore() {
+            public double getScore() {
                 return score;
             }
         };
